EXTREMELY IMPORTANT!! DEFAULT ACTIVATION CODE IS NOW 404 BY DEFAULT, NOT 200.
 - Added missing_status_code as configuration param too.

This if for a number of reasons (mainly for scalability for the web service),
but it also makes more sense that "404 Not Found" actually refers to the
"dissappeared" state of the computer. "200 OK" now means that the device is OK.

Syncronizing our Prey codes to HTTP responses will also allow us to make a more
semantic use of them (403 Forbidden, 302 Found, 409 Conflict, etc) which is also
nice.

The important thing is that standalone users of Prey will need to change the
setting in the config if they wish to keep waking Prey up by generating the URL
and not the other way around (which, in my opinion, is easier, since as its
already there there's no need to remember it or write it down).