Enter key on checkboxes, radioboxes and textboxes performs default action (next/apply)
Removed "manual" tying of apply_settings to the activate event of controls and control
the default button instead.

Corrected bad page forward/back calculation for http configuration, introduced when adding
the device screen.

If this is the first time configuration is run on a machine it is possible that the hostname
matches a current device title (e.g. reformat/reinstall of OS, using old values), so
hostname match is also checked to select a default device value in the dropdown.